extraction of notification to message
