+++ Turfy/Turfy/Message.swift
@@ -7,21 +7,43 @@
 //
 
 import Foundation
+
 import Firebase
+import UIKit
+import MapKit
+import CoreLocation
+
+struct GeoKey {
+    static let latitude = "latitude"
+    static let longitude = "longitude"
+    static let radius = "radius"
+    static let id = "id"
+    static let sender = "sender"
+    static let text = "message"
+    static let eventType = "eventType"
+}
+
+enum EventType: String {
+    case onEntry = "On Entry"
+    case onExit = "On Exit"
+}
 
 struct Message {
 	let id: String
 	let sender: String
 	let recipient: String
 	let text: String
-	let location: String
-	let radius: Int
+
+    let latitude: Double
+    let longitude: Double
+    let radius: CLLocationDistance
+    let eventType: EventType
 	let sentAt: String
 	let expires: Int
 	let dateformatter = DateFormatter()
 
 	
-	init(id: String, sender: String, recipient: String, location: String, text:String, radius: Int, expires: Int = 10) {
+    init(id: String, sender: String, recipient: String, text: String, latitude: Double, longitude: Double, radius: CLLocationDistance, eventType: EventType,   expires: Int = 10) {
 		
 		self.dateformatter.dateFormat = "dd/MM/yy h:mm"
 		let now = self.dateformatter.string(from: NSDate() as Date)
@@ -30,7 +52,7 @@ struct Message {
 		self.sender = sender
 		self.recipient = recipient
 		self.text = text
-		self.location = location
+
 		self.radius = radius
 		self.sentAt = now
 		self.expires = expires
@@ -43,7 +65,7 @@ struct Message {
             sender = snapshotValue["sender"] as! String
             recipient = snapshotValue["recipient"] as! String
             text = snapshotValue["text"] as! String
-            location = snapshotValue["location"] as! String
+
             radius = snapshotValue["radius"] as! Int
             sentAt = snapshotValue["sentAt"] as! String
             expires = snapshotValue["expires"] as! Int
@@ -56,7 +78,7 @@ struct Message {
             "sender": sender,
             "recipient": recipient,
             "text": text,
-            "location": location,
+
             "radius": radius,
             "sentAt": sentAt,
             "expires": expires,
